refactor: refactor song play ui-related code
