+++ data-struct-lib/src/song.rs
@@ -41,7 +41,7 @@ pub struct GameNote {
     #[serde(skip, default = "JangguFace::default")]
     pub face: JangguFace,
 }
-#[derive(Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct GameChart {
     pub artist: String,
     pub delay: u64,
@@ -241,10 +241,11 @@ impl GameSong {
                 .unwrap()
                 .to_string();
             if let Some(video_filename) = deserialized.video_filename {
-                deserialized.video_filename = Some(Path::join(path, video_filename)
-                    .to_str()
-                    .unwrap()
-                    .to_string()
+                deserialized.video_filename = Some(
+                    Path::join(path, video_filename)
+                        .to_str()
+                        .unwrap()
+                        .to_string(),
                 );
             }
             deserialized.cover_image_filename = Path::join(path, deserialized.cover_image_filename)
