+++ game/src/game/game_player/chart_player_ui/displayed_song_note.rs
@@ -1,0 +1,8 @@
+use bidrum_data_struct_lib::janggu::{JangguFace, JangguStick};
+
+#[derive(Clone)]
+pub struct DisplayedSongNote {
+    pub distance: f64,
+    pub face: JangguFace,
+    pub stick: JangguStick,
+}
