+++ game/src/game/tutorial/ending.rs
@@ -3,16 +3,12 @@ use std::{path, time::Instant};
 use kira::sound::static_sound::StaticSoundSettings;
 use sdl2::image::LoadTexture;
 
-use crate::game::{
-    game_common_context::GameCommonContext,
-    game_player::{draw_gameplay_ui::GamePlayUIResources, janggu_state_with_tick},
-};
+use crate::game::{game_common_context::GameCommonContext, game_player::janggu_state_with_tick};
 
 use super::display_tutorial_messages;
 
 pub(crate) fn do_tutorial_ending(
     common_context: &mut GameCommonContext,
-    game_ui_resources: &mut GamePlayUIResources,
     janggu_state_and_tutorial_start_time: &mut (
         &mut janggu_state_with_tick::JangguStateWithTick,
         Instant,
@@ -33,7 +29,6 @@ pub(crate) fn do_tutorial_ending(
 
     display_tutorial_messages(
         common_context,
-        game_ui_resources,
         &[message],
         janggu_state_and_tutorial_start_time,
     );
